Unify saving/importing/syncing configuration

+++ src/core/archive.py
@@ -1,10 +1,23 @@
 import os, sys, shutil, subprocess, argparse, re
 from savedesktop.globals import *
+from savedesktop.core.password_store import PasswordStore
 
 parser = argparse.ArgumentParser()
 parser.add_argument("-c", "--create", help="Create archive", type=str)
 parser.add_argument("-u", "--unpack", help="Unpack archive", type=str)
 args = parser.parse_args()
+
+TEMP_CACHE = f"{CACHE}/workspace"
+
+# Cleanup the cache dir before saving
+def cleanup_cache_dir():
+    print("Cleaning up the cache directory")
+    try:
+        shutil.rmtree(TEMP_CACHE)
+    except:
+        pass
+    os.makedirs(TEMP_CACHE, exist_ok=True)
+    os.chdir(TEMP_CACHE)
 
 # Get password entered in the "Create a new password" dialog from the temporary file
 def get_password():
@@ -13,26 +26,27 @@ def get_password():
         with open(temp_file) as tmp:
             return tmp.read().strip()
     else:
-        return None
+        p = PasswordStore()
+        return p.password
 
 # Remove above temporary file
 def remove_temp_file():
-    try:
+    if os.path.exists(f"{CACHE}/temp_file"):
         os.remove(f"{CACHE}/temp_file")
-    except FileNotFoundError:
-        pass
 
 class Create:
     def __init__(self):
         self.start_saving()
 
     def start_saving(self):
-        self._cleanup_cache_dir()
+        cleanup_cache_dir()
         subprocess.run([sys.executable, "-m", "savedesktop.core.config", "--save"], check=True, env={**os.environ, "PYTHONPATH": f"{app_prefix}"})
 
         print("Creating and moving the configuration archive or folder to the user-defined directory")
 
-        if settings["save-without-archive"]:
+        # In the periodic saving mode, it's not allowed to save the
+        # configuration without creating the archive
+        if settings["save-without-archive"] and not args.create == settings["periodic-saving-folder"]:
             self._copy_config_to_folder()
         else:
             self._create_archive()
@@ -41,25 +55,14 @@ class Create:
         print("Configuration saved successfully.")
         remove_temp_file()
 
-    # Cleanup the cache dir before saving
-    def _cleanup_cache_dir(self):
-        print("Cleaning up the cache directory")
-        save_cache_dir = f"{CACHE}/save_config"
-        try:
-            shutil.rmtree(save_cache_dir)
-        except:
-            pass
-        os.makedirs(save_cache_dir, exist_ok=True)
-        os.chdir(save_cache_dir)
-
     # Copy the configuration folder to the user-defined directory
     def _copy_config_to_folder(self):
-        open(f"{CACHE}/save_config/.folder.sd", "w").close()
+        open(f".folder.sd", "w").close()
 
         if os.path.exists(args.create):
             shutil.rmtree(args.create)
 
-        shutil.move(f"{CACHE}/save_config", f"{args.create}")
+        shutil.move(TEMP_CACHE, f"{args.create}")
 
     # Create a new ZIP archive with 7-Zip
     def _create_archive(self):
@@ -85,30 +88,14 @@ class Unpack:
         self.import_file = args.unpack
         self.import_folder = args.unpack
 
-        self._cleanup_cache_dir()
+        cleanup_cache_dir()
         self._check_config_type()
 
         self._replace_home_in_files(".", home)
         subprocess.run([sys.executable, "-m", "savedesktop.core.config", "--import_"], check=True, env={**os.environ, "PYTHONPATH": f"{app_prefix}"})
 
-        self._remove_status_file()
-
         print("Configuration imported successfully.")
         remove_temp_file()
-
-    def _cleanup_cache_dir(self):
-        # Cleanup the cache dir before importing
-        print("Cleaning up the cache directory")
-        imp_cache_dir = f"{CACHE}/import_config"
-        try:
-            shutil.rmtree(imp_cache_dir)
-        except:
-            pass
-        os.makedirs(imp_cache_dir, exist_ok=True)
-        os.chdir(imp_cache_dir)
-
-        # Create a txt file to prevent removing the cache's content after closing the app window
-        open("import_status", "w").close()
 
     # Check, if the input is archive or folder
     def _check_config_type(self):
@@ -130,7 +117,7 @@ class Unpack:
 
     # Copy the user-defined folder to the cache directory
     def _copy_folder_to_cache(self):
-        shutil.copytree(self.import_folder, f"{CACHE}/import_config", dirs_exist_ok=True, ignore_dangling_symlinks=True)
+        shutil.copytree(self.import_folder, TEMP_CACHE, dirs_exist_ok=True, ignore_dangling_symlinks=True)
 
     # Unpack the ZIP archive with 7-Zip
     def _unpack_zip_archive(self):
@@ -148,13 +135,13 @@ class Unpack:
             print("Checking password is completed.")
 
         subprocess.run(
-            ['7z', 'x', '-y', f'-p{password}', self.import_file, f'-o{CACHE}/import_config'],
+            ['7z', 'x', '-y', f'-p{password}', self.import_file, f'-o{TEMP_CACHE}'],
             capture_output=False, text=True, check=True
         )
 
     # Unpack a legacy archive with Tarball (for backward compatibility)
     def _unpack_tar_archive(self):
-        subprocess.run(["tar", "-xzf", self.import_file, "-C", f"{CACHE}/import_config"],capture_output=True, text=True, check=True)
+        subprocess.run(["tar", "-xzf", self.import_file, "-C", f"{TEMP_CACHE}"],capture_output=True, text=True, check=True)
 
     # Replace original /home/$USER path with actual path in the dconf-settings.ini file and other XML files
     def _replace_home_in_files(self, root, home, patterns=(".xml", ".ini")):
@@ -170,15 +157,6 @@ class Unpack:
                         with open(path, "w", encoding="utf-8") as f:
                             f.write(new_text)
                         print(f"Updated /home/$USER path in: {path}")
-
-    # Remove the "import_status" file if the condition is met
-    def _remove_status_file(self):
-        if all(not os.path.exists(p) for p in [
-            f"{CACHE}/import_config/app",
-            f"{CACHE}/import_config/installed_flatpaks.sh",
-            f"{CACHE}/import_config/installed_user_flatpaks.sh"
-        ]):
-            os.remove("import_status")
 
 if args.create:
     Create()
